// types/qat.ts
export interface Nomination {
	_id: string;
	type: EventType;
	timestamp: Timestamp;
	beatmapsetId: number;
	creatorId: number;
	creatorName: string;
	modes: GameModeName[];
	discussionId: number | null;
	userId: number;
	artistTitle: string;
	content: string | null;
	genre: string;
	language: string;
}

export interface NominationReset {
	_id: string;
	type: EventType;
	timestamp: Timestamp;
	beatmapsetId: number;
	creatorId: number;
	creatorName: string;
	modes: GameModeName[];
	discussionId: number | null;
	userId: number;
	content: string | null;
	genre: string;
	language: string;
	__v?: number;
	createdAt?: Timestamp;
	isBnOrNat?: boolean;
	isReviewed?: boolean;
	isUnique?: boolean;
	responsibleNominators?: [];
	updatedAt?: Timestamp;
	obviousness?: number | null;
	severity?: number | null;
}

export interface QACheck {
	_id: string;
	user: string;
	event: BeatmapEvent;
	timestamp: Timestamp;
	mode: string;
	createdAt: Timestamp;
	updatedAt: Timestamp;
	__v: number;
	comment?: string;
	id: string;
}

export interface DisqualifiedQACheck {
	modes: GameModeName[];
	obviousness: number | null;
	severity: number | null;
	isBnOrNat: boolean;
	isUnique: boolean;
	responsibleNominators: [];
	isReviewed: boolean;
	_id: string;
	type: EventType;
	timestamp: Timestamp;
	beatmapsetId: number;
	creatorId: number;
	creatorName: string;
	discussionId: number | null;
	userId: number | null;
	artistTitle: string;
	content: string | null;
	genre: string;
	language: string;
	__v: number;
	createdAt: Timestamp;
	updatedAt: Timestamp;
	qaComment?: string | null;
	id: string;
}

export interface UserActivity {
	uniqueNominations: Nomination[];
	nominationsDisqualified: NominationReset[]; // ? dqs recieved
	nominationsPopped: NominationReset[]; // ? pops recieved
	disqualifications: NominationReset[]; // ? dqs done by user
	pops: NominationReset[]; // ? pops done by user
	qualityAssuranceChecks: QACheck[];
	disqualifiedQualityAssuranceChecks: DisqualifiedQACheck[];
}

export interface UserActivityResponse {
	status: number;
	data: UserActivity;
}